requirements.txt
fastapi
uvicorn[standard]
flet==0.24.1
sqlmodel
sqlalchemy
pyjwt
pyotp
qrcode[pil]
ddgs
httpx
markdown2
pdfkit
python-dotenv
groq
google-generativeai
pydantic
websockets
loguru
aiofiles
jinja2
beautifulsoup4
